Finish register and memory support
In Linux, `sys/user.h` provides the basic register definition. We then
can use `ptrace(PTRACE_GETREGS, pid, nullptr, &regs)` to get all the
value of registers.

However, there are too many registers. So this commit uses scoped enum to
define the `Reg` and make an array to store it such that it can use STL
algorithm to avoid using tedious `switch` statements.

For memory, it is easy. Use `ptrace(PTRACE_PEEKDATA, ...)` for reading
and `ptrace(PTRACE_POKEDATA), ...` for writing.

For stepping over the breakpoint, it needs to get the current PC (`rip`)
register value, and make it substract 1 to get the address of the
breakpoint address. And set the PC to it. Last, this commit adds the
function `stepOverBreakpoint` to the function `continueExecution`.