Add source code display support
Linux use ELF and DWARF to store the specific information. This commit
use `libelfin` to parse the information for convenience.

This commit also changes the signal handling for elegance. When
receiving siganl, choose the specific siganl to handle, and wrap the PC
back operation. So the function `stepOverBreakpoint` can delete the PC
operation, and uses the function `waitForSignal` to solve this problem
elegantly.

Also, in the `handleSiganlTrap` function, this commit use `printSource`
function to print the current line of the program text.