Add source breakpoint support
For adding breakpoint at function, this commit iterates the compilation
units and find the subprogram's name attribute to get the address of the
function.

For adding breakpoint at line, this commit iterate the compilation units
line entry to find the address.

Finally, this commit add symbol lookup find ELF information.